LogisticRegressionTest contains error to clear
