+++ src/main/scala/io/github/jsarni/CaraStage/CaraStageMapper.scala
@@ -1,13 +1,12 @@
 package io.github.jsarni.CaraStage
 
 import io.github.jsarni.CaraStage.DatasetStage.CaraDataset
-import io.github.jsarni.CaraStage.ModelStage.{CaraModel, LogisticRegression, TestStage}
+import io.github.jsarni.CaraStage.ModelStage.{CaraModel, LogisticRegression}
 
 trait CaraStageMapper {
 
   def mapModelStage(stageDescription: CaraStageDescription): CaraModel = {
     stageDescription.stageName match {
-      case "TestStage" => TestStage(stageDescription.params)
       case "LogisticRegression" => LogisticRegression(stageDescription.params)
       case _ => throw
         new Exception(s"${stageDescription.stageName} is not a valid Model stage name. Please verify the corresponding Yaml File")
