+++ src/main/scala/io/github/jsarni/CaraStage/ModelStage/LogisticRegression.scala
@@ -2,6 +2,7 @@ package io.github.jsarni.CaraStage.ModelStage
 import io.github.jsarni.CaraStage.Annotation.MapperConstructor
 import org.apache.spark.ml.PipelineStage
 import org.apache.spark.ml.classification.{LogisticRegression => log}
+import scala.util.Try
 
 
 
@@ -31,7 +32,7 @@ case class LogisticRegression(MaxIter: Option[Int], RegParam: Option[Double], El
     )
   }
 
-  override def build(): PipelineStage = {
+  override def build(): Try[PipelineStage] = Try {
     val lr = new log()
     val definedFields = this.getClass.getDeclaredFields.filter(f => f.get(this).asInstanceOf[Option[Any]].isDefined)
     val names = definedFields.map(f => f.getName)
