--- src/main/scala/io/github/jsarni/CaraStage/ModelStage/TestStage.scala
@@ -1,25 +1,0 @@
-package io.github.jsarni.CaraStage.ModelStage
-
-import io.github.jsarni.CaraStage.Annotation.MapperConstructor
-import org.apache.spark.ml.PipelineStage
-
-
-case class TestStage(MaxIter: Option[Int], RegParam: Option[Double], ElasticNetParam: Option[Double])
-  extends CaraModel {
-
-  @MapperConstructor
-  def this(params: Map[String, String]) = {
-    this(
-      params.get("MaxIter").map(_.toInt),
-      params.get("RegParam").map(_.toDouble),
-      params.get("ElasticNetParam").map(_.toDouble)
-    )
-  }
-
-  override def build(): PipelineStage = ???
-}
-
-object TestStage {
-  def apply(params: Map[String, String]): TestStage = new TestStage(params)
-}
-
